+++ macos-leo/app/Components/Dock/DockItems.tsx
@@ -1,19 +1,18 @@
 import GlobalProfile from "@/public/assets/Icons/globalprofile.webp";
-import Laucnpad from "@/public/assets/Icons/launchpad.webp";
-import WebDiary from "@/public/assets/Icons/WebDiary.webp";
 import Safari from "@/public/assets/Icons/safari.webp";
 import Terminal from "@/public/assets/Icons/terminal.webp";
 import Typora from "@/public/assets/Icons/typora.webp";
 import Vscode from "@/public/assets/Icons/vscode.webp";
+import WebDiary from "@/public/assets/Icons/WebDiary.webp";
 import Youtube from "@/public/assets/Icons/YouTube.webp";
 import GloblaProfileApp from "../MainApps/GloblaProfileApp";
 import LiveTermMain from "../MainApps/LiveTerm/LiveTermMain";
 import SafariMain from "../MainApps/Safari/SafariMain";
 import TyporaApp from "../MainApps/TyporaApp";
 import VSCodeApp from "../MainApps/VSCodeApp";
+import WebDiaryApp from "../MainApps/WebDiary/WebDiaryApp";
 import YoutubeApp from "../MainApps/YoutubeApp";
 import { DockItem } from '../Types/FloatingDockType';
-import WebDiaryApp from "../MainApps/WebDiary/WebDiaryApp";
 
 
 export const DockItems: DockItem[] = [
@@ -22,8 +21,8 @@ export const DockItems: DockItem[] = [
     title: "WebDiary",
     imageSrc: WebDiary.src,
     content: <WebDiaryApp/>,
-    initialSize: { width: 800, height: 400 },
-    initialPosition: { x: 100, y: 100 },
+    initialSize: { width: 900, height: 470 },
+    initialPosition: { x: 200, y: 70 },
   },
   {
     id: "safari",