+++ macos-leo/app/Components/MainApps/WebDiary/components/WDMainSidebar.tsx
@@ -1,20 +1,28 @@
-import { IconAdjustments, IconBrandCupra, IconBrandGithub, IconCloudOff } from "@tabler/icons-react"
-import {WDMainSidebarProps} from "./types/WDType"
+import {
+  IconAdjustments,
+  IconBrandCupra,
+  IconBrandGithub,
+  IconCloudOff,
+} from "@tabler/icons-react";
+import { WDMainSidebarProps } from "./types/WDType";
 
-
-function WDMainSidebar({ items, selectedSection, onSectionSelect }: WDMainSidebarProps) {
+function WDMainSidebar({
+  items,
+  selectedSection,
+  onSectionSelect,
+}: WDMainSidebarProps) {
   const getIcon = (title: string) => {
     switch (title.toLowerCase()) {
-      case 'about':
+      case "about":
         return <IconBrandCupra size={20} />;
-      case 'projects':
+      case "projects":
         return <IconBrandGithub size={20} />;
       default:
         return null;
     }
   };
   return (
-    <div className="flex flex-col w-40 h-full bg-gray-500 text-white">
+    <div className="flex flex-col w-56 h-full bg-gray-700 text-white">
       <div className="flex justify-end gap-2 p-2">
         <IconCloudOff size={20} />
         <IconAdjustments size={20} />
@@ -23,8 +31,12 @@ function WDMainSidebar({ items, selectedSection, onSectionSelect }: WDMainSideba
         {items.map((item) => (
           <div
             key={item.id}
-            className={`flex justify-center items-center p-2 cursor-pointer
-              ${selectedSection?.id === item.id ? 'bg-orange-500' : 'hover:bg-gray-400'}`}
+            className={`flex px-4 items-center py-2 cursor-pointer
+              ${
+                selectedSection?.id === item.id
+                  ? "bg-orange-500"
+                  : "hover:bg-gray-400"
+              }`}
             onClick={() => onSectionSelect(item)}
           >
             {getIcon(item.title)}
@@ -33,7 +45,7 @@ function WDMainSidebar({ items, selectedSection, onSectionSelect }: WDMainSideba
         ))}
       </div>
     </div>
-  )
+  );
 }
 
-export default WDMainSidebar
+export default WDMainSidebar;