+++ macos-leo/app/Components/MainApps/WebDiary/components/WebDiaryItems.tsx
@@ -10,21 +10,21 @@ export const WDItemsData : WDData[] = [
                 title: "About me",
                 file: "web-diary",
                 smallDescription: "Web Diary is a simple diary app that allows you to write your thoughts and ideas in markdown format.",
-                link: "https://raw.githubusercontent.com/Renovamen/flint/main/README.md"
+                link: "https://raw.githubusercontent.com/Roshan-Here/Roshan-Here/master/README.md"
             },
             {
                 id: "2",
                 title: "Github staus",
                 file: "gitstatus",
                 smallDescription: "Web Diary is a simple diary app that allows you to write your thoughts and ideas in markdown format.",
-                link: "https://raw.githubusercontent.com/Renovamen/playground-macos/main/README.md"
+                link: "https://raw.githubusercontent.com/Roshan-Here/Roshan-Here/master/README.md"
             },
             {
                 id: "3",
                 title: "About this site",
                 file: "aboutsite",
                 smallDescription: "Web Diary is a simple diary app that allows you to write your thoughts and ideas in markdown format.",
-                link: "https://raw.githubusercontent.com/Renovamen/flint/main/README.md"
+                link: "https://raw.githubusercontent.com/Roshan-Here/Roshan-Here/master/README.md"
             }
             ]
         },
@@ -37,21 +37,21 @@ export const WDItemsData : WDData[] = [
                     title: "Project 1",
                     file: "proj1",
                     smallDescription: "Web Diary is a simple diary app that allows you to write your thoughts and ideas in markdown format.",
-                    link: "https://raw.githubusercontent.com/Renovamen/flint/main/README.md"
+                    link: "https://raw.githubusercontent.com/Roshan-Here/OneCompiler/refs/heads/development/Readme.md"
                 },
                 {
                     id: "2",
                     title: "Project 2",
                     file: "proj2",
                     smallDescription: "Web Diary is a simple diary app that allows you to write your thoughts and ideas in markdown format.",
-                    link: "https://raw.githubusercontent.com/Renovamen/flint/main/README.md"
+                    link: "https://raw.githubusercontent.com/Roshan-Here/Roshan-Here/master/README.md"
                 },
                 {
                     id: "3",
                     title: "Project 3",
                     file: "proj3",
                     smallDescription: "Web Diary is a simple diary app that allows you to write your thoughts and ideas in markdown format.",
-                    link: "https://raw.githubusercontent.com/Renovamen/flint/main/README.md"
+                    link: "https://raw.githubusercontent.com/Roshan-Here/Roshan-Here/master/README.md"
                 }
             ]
         }