+++ macos-leo/app/page.tsx
@@ -1,10 +1,10 @@
 "use client";
 
-import MorngBg from "@/public/assets/background/wallpaper-day.jpg";
-import NightBg from "@/public/assets/background/wallpaper-night.jpg";
+import MorngBg from "@/public/assets/background/wallpaper-day.webp";
+import NightBg from "@/public/assets/background/wallpaper-night.webp";
+import MainScreen from "./Components/MainScreen/MainScreen";
 import Navbar from "./Components/Navbar";
 import { UserThemeStore } from "./state/store";
-import MainScreen from "./Components/MainScreen/MainScreen";
 
 export default function Home() {
   const { isDark } = UserThemeStore();
@@ -14,15 +14,14 @@ export default function Home() {
       draggable={false}
       style={{
         backgroundImage: `url(${isDark ? NightBg.src : MorngBg.src})`,
-        backgroundSize: "cover", 
-        backgroundPosition: "center", 
-        backgroundRepeat: "no-repeat", 
+        backgroundSize: "cover",
+        backgroundPosition: "center",
+        backgroundRepeat: "no-repeat",
         filter: "brightness(106%)",
       }}
     >
       <Navbar />
-      <MainScreen/>
-
+      <MainScreen />
     </div>
   );
 }